refactor: move types to separate module
